backend/requirements.txt
fastapi>=0.104.0
uvicorn>=0.23.2
pydantic>=2.4.2
motor>=3.3.0  # MongoDB async driver
python-dotenv>=1.0.0
httpx>=0.27.0
python-jose>=3.3.0
passlib>=1.7.4
python-multipart>=0.0.6
pytest>=7.4.2
black>=23.9.1
flake8>=6.1.0
mypy>=1.6.1
isort>=5.12.0
openai>=1.2.0
langchain>=0.0.267
langchain-openai>=0.0.1
langgraph>=0.0.19
pydantic-settings
pydantic[email]
resend>=0.7.0
